+++ pendulum/tz/local_timezone.py
@@ -176,7 +176,6 @@ def _get_unix_timezone(_root='/'):  # type: (str) -> Timezone
     # OpenSUSE has a TIMEZONE setting in /etc/sysconfig/clock and
     # Gentoo has a TIMEZONE setting in /etc/conf.d/clock
     # We look through these files for a timezone:
-
     zone_re = re.compile('\s*ZONE\s*=\s*\"')
     timezone_re = re.compile('\s*TIMEZONE\s*=\s*\"')
     end_re = re.compile('\"')
@@ -185,6 +184,7 @@ def _get_unix_timezone(_root='/'):  # type: (str) -> Timezone
         tzpath = os.path.join(_root, filename)
         if not os.path.exists(tzpath):
             continue
+
         with open(tzpath, 'rt') as tzfile:
             data = tzfile.readlines()
 
@@ -194,13 +194,21 @@ def _get_unix_timezone(_root='/'):  # type: (str) -> Timezone
             if match is None:
                 # No ZONE= setting. Look for the TIMEZONE= setting.
                 match = timezone_re.match(line)
+
             if match is not None:
                 # Some setting existed
                 line = line[match.end():]
                 etctz = line[:end_re.search(line).start()]
 
-                # We found a timezone
-                return Timezone(etctz.replace(' ', '_'))
+                parts = etctz.replace(' ', '_').split('/')[-2:]
+                while parts:
+                    tzpath = '/'.join(parts)
+                    try:
+                        return Timezone(tzpath)
+                    except (ValueError, IOError, OSError):
+                        pass
+
+                    parts.pop(0)
 
     # systemd distributions use symlinks that include the zone name,
     # see manpage of localtime(5) and timedatectl(1)