+++ tests/tz/test_local_timezone.py
@@ -20,9 +20,22 @@ def test_unix_symlink():
     assert tz.name == 'Europe/Paris'
 
 
+@pytest.mark.skipif(sys.platform == 'win32',
+                    reason='Test only available for UNIX systems')
+def test_unix_clock():
+    # A ZONE setting in the target path of a symbolic linked localtime,
+    # f ex systemd distributions
+    local_path = os.path.join(os.path.split(__file__)[0], '..')
+    tz = _get_unix_timezone(
+        _root=os.path.join(local_path, 'fixtures', 'tz', 'clock')
+    )
+
+    assert tz.name == 'Europe/Zurich'
+
+
 @pytest.mark.skipif(sys.platform != 'win32',
                     reason='Test only available for Windows')
-def test_unix_symlink():
+def test_windows_timezone():
     timezone = _get_windows_timezone()
 
     assert timezone is not None